orders_service/requirements.txt
sqlalchemy==1.4.46
jose==1.0.0
fastapi==0.95.0
passlib==1.7.4
python-jose==3.3.0
psycopg2-binary==2.9.6
uvicorn==0.21.1
pydantic==1.10.7
pydantic[email]==1.10.7
jinja2==3.0.3
kafka-python==2.0.2
redis==4.5.4
requests==2.28.1
strawberry-graphql==0.135.0
reportlab==4.2.5
